#COBUMMFO-81# Коментар: MERGE RC_42_RU to MASTER

+++ ABS/master/web/ru/barsroot/App_Code/Areas/CreditUi/Models/UpdateProvide.cs
@@ -29,7 +29,7 @@ namespace BarsWeb.Areas.CreditUi.Models
 
         public decimal? DEL { get; set; }
 
-        public int? DEPID { get; set; }
+        public decimal? DEPID { get; set; }
 
         public decimal? SV { get; set; }
